+++ labs_god/lab4/choice_sort/choice_funcs.h
@@ -2,8 +2,8 @@
 #define CHOICE_FUNCS_H
 
 // Прототипы. Реализации лежат в сhoice_funcs.c
-int findMin(int array[], int size, int start, int *compares);
+int findMin(int array[], int size, int start, long long *compares);
 void Swap(int array[], int i, int j);
-int* choiceSort(int array[], int size, int *compares, int *swaps);
+int* choiceSort(int array[], int size, long long *compares, long long *swaps);
 
 #endif
